+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 * Add missing `provider` option for `urlForAuthentication`
+* Job status improvements
 
 ### 5.10.2 / 2021-12-03
 * Fix Scheduler and Component types