+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-### Unreleased
+### 5.10.3 / 2021-12-16
 * Add missing `provider` option for `urlForAuthentication`
 * Job status improvements